+++ cookbooks/third-party/selinux-6.1.12/test/cookbooks/selinux_test/metadata.rb
@@ -1,5 +1,0 @@
-name 'selinux_test'
-version '0.0.1'
-
-depends 'selinux'
-depends 'apt'