+++ cookbooks/third-party/selinux-6.1.12/test/cookbooks/selinux_test/recipes/permissive_resource.rb
@@ -1,3 +1,0 @@
-selinux_permissive 'httpd_t'
-
-selinux_permissive 'user_t'