+++ cookbooks/third-party/selinux-6.1.12/test/integration/permissive/controls/module_control.rb
@@ -1,9 +1,0 @@
-control 'module' do
-  title 'Verify that SELinux modules are installed correctly'
-
-  describe selinux.modules.where(name: 'test') do
-    it { should exist }
-    it { should be_installed }
-    it { should be_enabled }
-  end
-end