+++ CHANGELOG.md
@@ -13,6 +13,10 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 **CHANGES**
 - Upgrade Cinc Client to version to 18.4.12 from 18.2.7.
 - Upgrade Slurm to 23.11.7 (from 23.11.4).
+- Upgrade third-party cookbook dependencies:
+  - apt-7.5.22 (from apt-7.5.14)
+  - openssh-2.11.12 (from openssh-2.11.3)
+  - selinux-6.1.18 (from selinux-6.1.12)
 
 **BUG FIXES**
 - Fixed an issue that prevented cluster updates from including EFS filesystems with encryption in transit.